Fix “get_posts() is discouraged in favor of creating a new WP_Query() so that Advanced Post Cache will cache the query”.

+++ classes/Pronamic/WP/Pay/Plugin.php
@@ -969,7 +969,6 @@ class Pronamic_WP_Pay_Plugin {
 		);
 
 		$args = array(
-			'fields'      => 'ids',
 			'post_type'   => 'pronamic_pay_subscr',
 			'nopaging'    => true,
 			'orderby'     => 'post_date',
@@ -995,44 +994,52 @@ class Pronamic_WP_Pay_Plugin {
 			),
 		);
 
-		$subscriptions = get_posts( $args );
+		$query = new WP_Query( $args );
 
-		foreach ( $subscriptions as $subscription_id ) {
-			$subscription = new Pronamic_WP_Pay_Subscription( $subscription_id );
-			$first        = $subscription->get_first_payment();
-			$gateway      = Pronamic_WP_Pay_Plugin::get_gateway( $first->config_id );
+		if ( $query->have_posts() ) {
+			while ( $query->have_posts() ) {
+				$query->the_post();
 
-			$payment = self::start_recurring( $subscription, $gateway );
+				$subscription_id = get_the_ID();
 
-			if ( $payment ) {
-				$frequency = $subscription->get_frequency();
-				$schedule  = false;
+				$subscription = new Pronamic_WP_Pay_Subscription( $subscription_id );
+				$first        = $subscription->get_first_payment();
+				$gateway      = Pronamic_WP_Pay_Plugin::get_gateway( $first->config_id );
 
-				if ( '' === $frequency ) {
-					$schedule = true;
-				}
+				$payment = self::start_recurring( $subscription, $gateway );
 
-				$next = $subscription->get_next_payment_datetime( 1 );
+				if ( $payment ) {
+					$frequency = $subscription->get_frequency();
+					$schedule  = false;
 
-				if ( ! $schedule ) {
-					$final = $subscription->get_final_payment_datetime();
-
-					if ( $final && $next < $final ) {
+					if ( '' === $frequency ) {
 						$schedule = true;
 					}
+
+					$next = $subscription->get_next_payment_datetime( 1 );
+
+					if ( ! $schedule ) {
+						$final = $subscription->get_final_payment_datetime();
+
+						if ( $final && $next < $final ) {
+							$schedule = true;
+						}
+					}
+
+					if ( $schedule ) {
+						update_post_meta( $subscription->get_id(), '_pronamic_subscription_next_payment', $next->format( 'Y-m-d H:i:s' ) );
+					} else {
+						wp_schedule_single_event( $final->getTimestamp(), 'pronamic_pay_subscription_completed', array( $subscription_id ) );
+
+						delete_post_meta( $subscription->get_id(), '_pronamic_subscription_next_payment' );
+						delete_post_meta( $subscription->get_id(), '_pronamic_subscription_renewal_notice' );
+					}
+
+					self::update_payment( $payment, false );
 				}
-
-				if ( $schedule ) {
-					update_post_meta( $subscription->get_id(), '_pronamic_subscription_next_payment', $next->format( 'Y-m-d H:i:s' ) );
-				} else {
-					wp_schedule_single_event( $final->getTimestamp(), 'pronamic_pay_subscription_completed', array( $subscription_id ) );
-
-					delete_post_meta( $subscription->get_id(), '_pronamic_subscription_next_payment' );
-					delete_post_meta( $subscription->get_id(), '_pronamic_subscription_renewal_notice' );
-				}
-
-				self::update_payment( $payment, false );
 			}
+
+			wp_reset_postdata();
 		}
 
 		self::send_subscription_renewal_notices();
@@ -1064,23 +1071,31 @@ class Pronamic_WP_Pay_Plugin {
 			),
 		);
 
-		$subscriptions = get_posts( $args );
+		$query = new WP_Query( $args );
 
-		foreach ( $subscriptions as $subscription_id ) {
-			$subscription = new Pronamic_WP_Pay_Subscription( $subscription_id );
+		if ( $query->have_posts() ) {
+			while ( $query->have_posts() ) {
+				$query->the_post();
 
-			do_action( 'pronamic_subscription_renewal_notice_' . $subscription->get_source(), $subscription );
+				$subscription_id = get_the_ID();
 
-			// Set next renewal date meta
-			$next_renewal = $subscription->get_next_payment_datetime( 1 );
-			$next_renewal->modify( '-1 week' );
+				$subscription = new Pronamic_WP_Pay_Subscription( $subscription_id );
 
-			// Prevent multiple notices being send for the same renewal date
-			if ( $next_renewal < $subscription->get_next_payment_datetime() ) {
-				$next_renewal = $subscription->get_next_payment_datetime();
+				do_action( 'pronamic_subscription_renewal_notice_' . $subscription->get_source(), $subscription );
+
+				// Set next renewal date meta
+				$next_renewal = $subscription->get_next_payment_datetime( 1 );
+				$next_renewal->modify( '-1 week' );
+
+				// Prevent multiple notices being send for the same renewal date
+				if ( $next_renewal < $subscription->get_next_payment_datetime() ) {
+					$next_renewal = $subscription->get_next_payment_datetime();
+				}
+
+				update_post_meta( $subscription_id, '_pronamic_subscription_renewal_notice', $next_renewal->format( 'Y-m-d H:i:s' ) );
 			}
 
-			update_post_meta( $subscription_id, '_pronamic_subscription_renewal_notice', $next_renewal->format( 'Y-m-d H:i:s' ) );
+			wp_reset_postdata();
 		}
 	}
